+++ src/main/java/org/example/iw_payment_service/dto/PaymentResponse.java
@@ -1,11 +1,10 @@
 package org.example.iw_payment_service.dto;
 
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.example.iw_payment_service.model.enums.PaymentStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 @Data
 @AllArgsConstructor
@@ -15,5 +14,5 @@ public class PaymentResponse {
     private Long userId;
     private LocalDateTime timestamp;
     private PaymentStatus status;
-    private Double paymentAmount;
+    private BigDecimal paymentAmount;
 }
